+++ R/epiweek.R
@@ -116,8 +116,9 @@ epiweek <- function(year = integer(), week = integer()) {
     }
 
     # check compatible lengths
-    if (length(year) != length(week))
-        stop("`year` and `week` must be the same length.")
+    tmp <- .recycle(year, week)
+    year <- tmp[[1L]]
+    week <- tmp[[2L]]
 
     .epiweek(year = year, week = week)
 }