+++ tests/testthat/test-epiweek.R
@@ -299,7 +299,8 @@ test_that("epiweek, miscellaneous work", {
     )
 
     expect_identical(epiweek(1L,1L),epiweek(1.5,1.5))
-    expect_error(epiweek(1L), "`year` and `week` must be the same length.", fixed = TRUE)
+    expect_error(epiweek(1L), "Cannot recycle a vector of length 0:", fixed = TRUE)
+    expect_error(epiweek(1:2,1:3), "Can only recycle vectors of length 1:", fixed = TRUE)
     expect_error(epiweek(year = character()), "`year` must be integer.", fixed = TRUE)
     expect_error(epiweek(week = character()), "`week` must be integer.", fixed = TRUE)
 })