+++ NEWS.md
@@ -1,5 +1,8 @@
 # grates (development version)
 
+* Constructors `yearweek()`, `epiweek()`, `isoweek()`, `yearmonth()` and
+  `yearquarter()` now recycle inputs of length 1.
+
 # grates 1.0.0
 
 For the first major release of grates a significant refactor has been undertaken